Remove recursive behavior from transform function
The transform function within a class extended from RewriteRule will not need to be recursive -- the Rule Transformer that uses the defined rules will map over the sequence of nodes itself -- no recursion necessary.

Including the recursive behavior, when used with the intended Rule Transformer will result in each node in the sequence getting a recurisve call on it. This effectively makes the function increase in complexity exponentially, depending on how deep the nesting of the XML/SVG nodes are.