Add support for namespaced attributes in FillTransformer
Namespaced attributes of Nodes are accessed with the same-named function (but with different arity) from their non-namespaced counterparts.
This necessitates one of two options: a reduntant namespace check for each node that gets fed to the FillTransformer, or a separate namespaced-subtype.

This commit will include this namespaced-handling subtype as well as a factory that will decide which type of FillTransformer to instantiate based on its arguments.

The full implementation of the factory's apply method is still buggy; the NamespacedFillTransformer's CheckID method is still problematic as well.