Refactor checkID method
previous implementations of the checkID used .attribute(ATTRIBUTE) syntax to receive attributes. This method would return an option of a sequence of a node, which would have to require chained methods to access the value nested within the list and convert it into text. If the option returns as type None instead of Some, the method will generate a runtime error of accessing the head of an empty list.

Instead of implementing checks for None / Some, checkID now uses the Node \ "@ATTRIBUTE" syntax and uses string interpolation.